REFACTOR: useSocket hook을 모듈로 간소화
